Profile Half Working & Customized Scroll

+++ src/app/app.component.ts
@@ -25,14 +25,17 @@ export class AppComponent implements OnInit, OnDestroy {
   }
 
   ngOnInit(): void {
-    this.crudService.GETWithAuth(api.getRefreshJwtURL(), 'refresh')
-      .subscribe(response => {
-        sessionStorage.setItem('User-Auth', JSON.stringify(response['auth:user'].user));
-        sessionStorage.setItem('X-Authentication-JWT', response['X-Authentication-JWT']);
-        sessionStorage.setItem('X-Encode-ID', response['X-Encode-ID']);
-        localStorage.setItem('X-Refresh-JWT', response['X-Refresh-JWT']);
-        this.router.navigate(['/video/categories']).then();
-      }, () => localStorage.clear());
+    if (!sessionStorage.getItem('User-Auth') && !sessionStorage.getItem('X-Authentication-JWT') &&
+      !sessionStorage.getItem('X-Encode-ID')) {
+      this.crudService.GETWithAuth(api.getRefreshJwtURL(), 'refresh')
+        .subscribe(response => {
+          sessionStorage.setItem('User-Auth', JSON.stringify(response['auth:user'].user));
+          sessionStorage.setItem('X-Authentication-JWT', response['X-Authentication-JWT']);
+          sessionStorage.setItem('X-Encode-ID', response['X-Encode-ID']);
+          localStorage.setItem('X-Refresh-JWT', response['X-Refresh-JWT']);
+          this.router.navigate(['/video/categories']).then();
+        }, () => localStorage.clear());
+    }
   }
 
   ngOnDestroy(): void {
